Reified integer <-> set constraints
git-svn-id: svn+ssh://svn.gecode.org/srv/gecode/svn/gecode/trunk@15566 64335634-5103-0410-b293-fc3d331e086d

+++ gecode/set/cardinality.cpp
@@ -74,7 +74,7 @@ namespace Gecode {
 
   void
   cardinality(Home home, SetVar s, IntVar x, Reify r) {
-    IntVar y(home, 0, Set::Limits::card);
+    IntVar y(home, 0, static_cast<int>(Set::Limits::card));
     rel(home, x, IRT_EQ, y, r);
     cardinality(home, s, y);
   }
